SP-3: Add PostController#likePost method and implement supporting methods

+++ src/main/java/org/practicum/yandex/service/post/PostService.java
@@ -20,7 +20,7 @@ public interface PostService {
 
     void deletePost(final BigInteger postId) throws DataNotFoundException;
 
-    int incrementLikes(final BigInteger postId);
+    BigInteger incrementLikes(final BigInteger postId);
 
     void updatePostImage(byte[] content);
